Implement skeleton for altitude controller

+++ imu.h
@@ -4,6 +4,6 @@
 void initIMU();
 void readIMU();
 void getGyro(double*, double*, double*);
-void getAcc(short*, short*, short*);
+void getAcc(double*, double*, double*);
 
 #endif
